# Conflicts:
#	src/main/java/seng202/team5/gui/DataListPageController.java
#	src/main/java/seng202/team5/gui/DetailedViewPageController.java
#	src/main/java/seng202/team5/repository/WineDAO.java
#	src/main/resources/fxml/DataListPage.fxml
#	src/main/resources/fxml/DetailedViewPage.fxml
#	src/test/java/seng202/team5/cucumber/SearchFilterStepDefs.java
#	src/test/java/seng202/team5/unittests/services/DataLoadServiceTest.java

+++ src/main/java/seng202/team5/gui/DetailedViewPageController.java
@@ -8,10 +8,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Random;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.geometry.Insets;
 import javafx.scene.Cursor;
 import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextArea;
@@ -19,8 +23,11 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.FlowPane;
+import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.controlsfx.control.PopOver;
@@ -32,6 +39,7 @@ import seng202.team5.models.Wine;
 import seng202.team5.repository.AssignedTagsDAO;
 import seng202.team5.repository.ReviewDAO;
 import seng202.team5.repository.TagsDAO;
+import seng202.team5.services.OpenWindowsService;
 import seng202.team5.services.UserService;
 import seng202.team5.services.WineService;
 
@@ -43,17 +51,20 @@ import java.util.Objects;
  *
  * @author Finn Brown
  */
-public class DetailedViewPageController extends PageController {
+public class DetailedViewPageController extends PageController implements ClosableWindow {
     private static final Logger log = LogManager.getLogger(DetailedViewPageController.class);
     private final Image emptyStar = new Image(
             Objects.requireNonNull(getClass().getResourceAsStream("/images/empty_star.png")));
     private final Image filledStar = new Image(
             Objects.requireNonNull(getClass().getResourceAsStream("/images/filled_star.png")));
+
     Label addLabel;
     boolean canAddTag = true;
     TagsDAO tagsDAO;
     AssignedTagsDAO assignedTagsDAO;
     List<Tag> tagsList;
+    @FXML
+    public Label ratingLogInLabel;
     @FXML
     private Button backButton;
     @FXML
@@ -84,6 +95,8 @@ public class DetailedViewPageController extends PageController {
     private ImageView wineColourImage;
 
     @FXML
+    private Label vineyardLabel;
+    @FXML
     private TextArea notesTextArea;
     @FXML
     private Button saveNotesButton;
@@ -101,6 +114,10 @@ public class DetailedViewPageController extends PageController {
     private ImageView star5;
     @FXML
     private FlowPane tagBox;
+
+    @FXML
+    private GridPane headerGridPane;
+
     private int selectedWineId;
     private int userId;
     private PopOver tagPopover;
@@ -113,11 +130,27 @@ public class DetailedViewPageController extends PageController {
      */
     @FXML
     private void initialize() {
+        OpenWindowsService.getInstance().addWindow(this);
+
         Wine selectedWine = WineService.getInstance().getSelectedWine();
         selectedWineId = selectedWine.getId();
         reviewDAO = new ReviewDAO();
         review = null;
 
+
+        initWineInfo(selectedWine);
+        initUserReviews();
+        initAdminActions();
+
+
+    }
+
+    /**
+     * Init all the labels with the wine information of the selected wine.
+     *
+     * @param selectedWine the selected wine object
+     */
+    private void initWineInfo(Wine selectedWine) {
         if (selectedWine != null) {
             nameLabel.setText(selectedWine.getName());
             priceLabel.setText("Price: $" + selectedWine.getPrice());
@@ -128,8 +161,14 @@ public class DetailedViewPageController extends PageController {
             varietyLabel.setText("Variety: " + selectedWine.getWineVariety());
             colourLabel.setText("Colour: " + selectedWine.getWineColour());
             setColourImage(selectedWine);
+            vineyardLabel.setText("Vineyard: " + selectedWine.getVineyard().getName());
         }
+    }
 
+    /**
+     * Initialise user specific review info and tags if applicable.
+     */
+    private void initUserReviews() {
         if (UserService.getInstance().getCurrentUser() != null) {
             // Get the user ID
             userId = UserService.getInstance().getCurrentUser().getId();
@@ -154,7 +193,9 @@ public class DetailedViewPageController extends PageController {
             }
 
             // Done Loading Tags
+            ratingLogInLabel.setVisible(false);
             logInMessageLabel.setText("");
+            ratingStars.setVisible(true);
             addTagLabel.setText("");
             if (review != null) {
                 notesTextArea.setText(review.getNotes());
@@ -167,6 +208,8 @@ public class DetailedViewPageController extends PageController {
             notesTextArea.setEditable(true);
         } else {
             logInMessageLabel.setText("Log in to save your notes!");
+            ratingLogInLabel.setVisible(true);
+            ratingStars.setVisible(false);
             addTagLabel.setText("Log in to add tags!");
             favoriteToggleButton.setDisable(true);
             saveNotesButton.setDisable(true);
@@ -185,6 +228,48 @@ public class DetailedViewPageController extends PageController {
             case "White" -> wineColourImage.setImage(new Image(Objects.requireNonNull(this.getClass().getResourceAsStream("/images/whiteColourWine.png"))));
             case "Rosé" -> wineColourImage.setImage(new Image(Objects.requireNonNull(this.getClass().getResourceAsStream("/images/roseColourWine.png"))));
         }//otherwise keep default image (unknown colour)
+    }
+
+    /**
+     * Initialise the button for the admin to edit a wine.
+     */
+    private void initAdminActions() {
+        if (UserService.getInstance().getCurrentUser() != null
+                && UserService.getInstance().getCurrentUser().getIsAdmin()) {
+            Button editWineButton = new Button("Edit Wine");
+            editWineButton.getStyleClass().add("detailed_view");
+            editWineButton.applyCss();
+            editWineButton.setOnAction(this::editWine);
+            headerGridPane.add(editWineButton, 1, 0);
+            GridPane.setMargin(editWineButton, new Insets(0, 10, 10, 0));
+        }
+    }
+
+    /**
+     * Called when the admin clicks the edit wine button.
+     *
+     * @param event the ActionEvent associated with the button click
+     */
+    private void editWine(ActionEvent event) {
+        Wine selectedWine = WineService.getInstance().getSelectedWine();
+        try {
+            FXMLLoader editWineLoader = new FXMLLoader(getClass()
+                    .getResource("/fxml/EditWinePopup.fxml"));
+            Parent root = editWineLoader.load();
+            Scene scene = new Scene(root);
+            Stage stage = new Stage();
+            stage.setScene(scene);
+            stage.setMinHeight(486);
+            stage.setMinWidth(762);
+            stage.setTitle(String.format("Edit wine %s", selectedWine.getName()));
+            String styleSheetUrl = MainWindow.styleSheet;
+            scene.getStylesheets().add(styleSheetUrl);
+            stage.initModality(Modality.WINDOW_MODAL);
+            stage.showAndWait();
+            initWineInfo(selectedWine);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     /**
@@ -459,26 +544,33 @@ public class DetailedViewPageController extends PageController {
      * Closes the page.
      */
     @FXML
-    private void handleBackButtonAction() {
+    @Override
+    public void closeWindow() {
+        OpenWindowsService.getInstance().closeWindow(this);
         close();
     }
 
     /**
-     * Closes the page.
+     * Closes the page and updates the review if necessary.
      */
-    private void close() {
+    public void close() {
+
         try {
             if (UserService.getInstance().getCurrentUser() != null) {
-                assignedTagsDAO.deleteFromUserWineId(userId, selectedWineId);
+                if (assignedTagsDAO != null) {
 
-                // Add review to this wine
-                if (!tagsList.isEmpty()) {
-                    createReviewIfNotExists();
-                }
+                    assignedTagsDAO.deleteFromUserWineId(userId, selectedWineId);
 
-                // Add to assigned tags db
-                for (Tag tag : tagsList) {
-                    assignedTagsDAO.add(new AssignedTag(tag.getTagId(), userId, selectedWineId));
+                    // Add review to this wine
+                    if (!tagsList.isEmpty()) {
+                        createReviewIfNotExists();
+                    }
+
+                    // Add to assigned tags db
+                    for (Tag tag : tagsList) {
+                        assignedTagsDAO.add(new AssignedTag(tag.getTagId(),
+                                userId, selectedWineId));
+                    }
                 }
             }
         } catch (DuplicateEntryException e) {
@@ -491,6 +583,7 @@ public class DetailedViewPageController extends PageController {
             addNotification("Updated Wine Review", "#d5e958");
         }
     }
+
 
 
     /**
